refactor(types): update property detail types and imports for better structure and clarity
