+++ apps/web/src/app/properties/[slug]/components/AmenitiesSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { AmenityIcon } from "./AmenityIcon";
-import type { Property } from "./types";
+import type { Property, Facility } from "@/types/property-detail";
 
 export function AmenitiesSection({ property }: { property: Property }) {
   if (!property.facilities?.length) return null;
@@ -9,7 +9,7 @@ export function AmenitiesSection({ property }: { property: Property }) {
     <section id="amenities" className="mt-6">
       <h3 className="text-xl font-semibold mb-6">Amenities & Facilities</h3>
       <div className="grid grid-cols-2 md:grid-cols-3 gap-4">
-        {property.facilities?.map((facility) => (
+        {property.facilities?.map((facility: Facility) => (
           <div key={facility.id} className="flex items-center gap-2">
             <AmenityIcon amenity={facility.facility} />
           </div>
